Enabling cmd.exe extensions for android.bat
Since Windows XP (at least?), cmd.exe has a bunch of extensions
that are enabled on Windows by default. However, some systems can
disable this globally in the registry. In that case, our
android.bat file will fail to run, because it relies on extended
features.

The solution is to explicitly turn on extensions via the
setlocal command, so at least it's turned on for our script. Most
users won't be affected by this (because they should already have
extensions turned on by default), but for the few users who have
disabled it for whatever reason, this will fix their case.
Otherwise, we'd just break with strange, unrelated failures.

https://code.google.com/p/android/issues/detail?id=78654

Change-Id: Ib5d651fbc7243c6b91f68b76d9da39f3a8445211