Quote xcopy arguments to fix "Invalid number of parameters" error
On Windows, android.bat calls xcopy with %TEMP% as part of its
arguments. If android.bat is called from MSYS (like Git Bash) %TEMP%
contains forward slashes, confusing xcopy which is trying to take the
path as command line arguments. This causes the "Invalid number of
parameters" error message on the console. Fix this by quoting the path
arguments to xcopy.

Change-Id: I24b75b44610b539614ea42f26d702eb3da9514e6